Origin: appearance fixes for osrs decoding

+++ src/main/java/com/cryptic/model/entity/player/save/PlayerSave.java
@@ -870,7 +870,7 @@ public class PlayerSave {
         private final Item topPkerReward;
         private final boolean female;
         private final int[] looks;
-        private final int[] colors;
+        private final short[] colors;
 
         //Combat attribs
         private final String spellBook;
